completed test for form fields
