+++ client/src/components/App.jsx
@@ -49,8 +49,7 @@ export default function App() {
   useEffect(
     () => {
       if (user) {
-          axios
-            .get(`https://www.googleapis.com/oauth2/v1/userinfo?access_token=${user.access_token}`, {
+          axios.get(`https://www.googleapis.com/oauth2/v1/userinfo?access_token=${user.access_token}`, {
                 headers: {
                   Authorization: `Bearer ${user.access_token}`,
                   Accept: 'application/json'
